Map permissions, player rating params, TU1 options
Added code to read a game variant's map permissions, player rating params, option toggles, and TU1 options. (This includes a member function on cobb::bitstream to read the compressed floats that Reach uses for TU1 options.) We now have parity with the JavaScript prototype.

Brought in cobb::bitset to make it easier to load option toggles.

+++ native/src/ReachVariantTool/formats/localized_string_table.cpp
@@ -66,7 +66,7 @@ void* ReachStringTable::_make_buffer(cobb::bitstream& stream) const noexcept {
 void ReachStringTable::read(cobb::bitstream& stream) noexcept {
    size_t count = stream.read_bits(this->count_bitlength);
    this->strings.resize(count);
-   for (int i = 0; i < count; i++)
+   for (size_t i = 0; i < count; i++)
       this->strings[i].read_offsets(stream, *this);
    if (count) {
       auto buffer = this->_make_buffer(stream);
